dashboard: Move dashboard into its own component
Closes #2756
Closes #2690
