+++ pkg/dashboard/manifest.json
@@ -2,7 +2,7 @@
     "version": 0,
 
     "dashboard": {
-        "_main_": {
+        "list": {
             "path": "dashboard/list",
             "label": "Dashboard",
             "order": 0
