+++ pkg/shell/index.html
@@ -60,7 +60,7 @@
             <a id="machine-color"></a>
             <a id="machine-link" class="dropdown-toggle" data-toggle="dropdown"
                 aria-haspopup="true" role="button" aria-expanded="false">
-               <img src="images/server-small.png" id="machine-avatar" class="machine-avatar"><span>Machines</span>
+               <img src="../shell/images/server-small.png" id="machine-avatar" class="machine-avatar"><span>Machines</span>
                <b class="caret"></b>
             </a>
             <ul class="dropdown-menu" role="menu" aria-labelledby="machine-link">
