+++ po/POTFILES.in
@@ -1,6 +1,8 @@
 [encoding: UTF-8]
 # List of source files containing translatable strings.
 # Please keep this file sorted alphabetically.
+[type: gettext/glade] pkg/dashboard/list.html
+pkg/dashboard/list.js
 [type: gettext/glade] pkg/docker/containers.html
 [type: gettext/glade] pkg/docker/console.html
 pkg/docker/containers.js
@@ -25,12 +27,10 @@ pkg/realmd/operation.js
 [type: gettext/glade] pkg/shell/shell.html
 pkg/shell/authorized-keys.js
 pkg/shell/cockpit-cpu-status.js
-pkg/shell/cockpit-dashboard.js
 pkg/shell/cockpit-main.js
 pkg/shell/cockpit-memory-status.js
 pkg/shell/cockpit-plot.js
 pkg/shell/cockpit-server.js
-pkg/shell/cockpit-setup.js
 pkg/shell/cockpit-util.js
 [type: gettext/glade] pkg/shell/index.html
 pkg/shell/index.js
