+++ test/check-dashboard
@@ -88,7 +88,7 @@ class TestDashboard(MachineCase):
         m3.start()
         m3.wait_boot()
 
-        self.login_and_go("dashboard", href="/dashboard/list", host=None)
+        self.login_and_go("/dashboard/list", host=None)
         self.inject_extras(b)
 
         self.wait_dashboard_addresses (b, [ "localhost" ])
@@ -96,7 +96,7 @@ class TestDashboard(MachineCase):
         # Start second browser and check that it is in sync
         b2 = self.new_browser()
         b2.default_user = "root"
-        b2.login_and_go("dashboard", href="/dashboard/list", host=None)
+        b2.login_and_go("/dashboard/list", host=None)
         self.inject_extras(b2)
         self.wait_dashboard_addresses (b2, [ "localhost" ])
         b.wait_present("#dashboard-hosts a:nth-child(1) button.pficon-delete.disabled")
@@ -153,7 +153,7 @@ class TestDashboard(MachineCase):
         b = self.browser
         m = self.machine
 
-        self.login_and_go("dashboard", href="/dashboard/list", host=None)
+        self.login_and_go("/dashboard/list", host=None)
         self.inject_extras(b)
         self.wait_dashboard_addresses (b, [ "localhost" ])
         old_style = b.attr("#dashboard-hosts .list-group-item[data-address='localhost']", "style")
@@ -205,7 +205,7 @@ class TestDashboard(MachineCase):
 
         # Sync them via Setup.
 
-        self.login_and_go("dashboard", href="/dashboard/list", host=None)
+        self.login_and_go("/dashboard/list", host=None)
         self.inject_extras(b)
 
         self.wait_dashboard_addresses (b, [ "localhost" ])
