+++ test/check-multi-machine
@@ -90,7 +90,7 @@ class TestMultiMachine(MachineCase):
         m2 = self.machine2
         m3 = self.machine3
 
-        self.login_and_go("dashboard", href="dashboard/list", host=None)
+        self.login_and_go("/dashboard/list", host=None)
         inject_extras(b)
         wait_dashboard_addresses (b, [ "localhost" ])
 
@@ -113,21 +113,21 @@ class TestMultiMachine(MachineCase):
         # https://github.com/cockpit-project/cockpit/issues/606
 
         kill_user_admin(m2)
-        b.wait_attr('#dashboard-hosts a:nth-child(%d) img' % m2_index, 'src', 'images/server-error.png')
+        b.wait_attr('#dashboard-hosts a:nth-child(%d) img' % m2_index, 'src', '../shell/images/server-error.png')
 
         kill_user_admin(m3)
-        b.wait_attr('#dashboard-hosts a:nth-child(%d) img' % m3_index, 'src', 'images/server-error.png')
+        b.wait_attr('#dashboard-hosts a:nth-child(%d) img' % m3_index, 'src', '../shell/images/server-error.png')
 
         # Navigating reconnects
-        b.click('#dashboard-hosts a:nth-child(%d)' % m2_index)
+        b.click("#dashboard-hosts a[data-address='%s']" % m2.address)
         b.switch_to_top()
         b.wait_js_cond('window.location.hash != "#/dashboard/list"')
         b.enter_page("server", m2.address)
         b.wait_text_not("#system_information_hostname_button", "")
         b.switch_to_top()
         b.go("/dashboard/list")
-        b.enter_page("dashboard")
-        b.wait_attr('#dashboard-hosts a:nth-child(%d) img' % m2_index, 'src', 'images/server-small.png')
+        b.enter_page("/dashboard/list")
+        b.wait_attr('#dashboard-hosts a:nth-child(%d) img' % m2_index, 'src', '../shell/images/server-small.png')
 
         b.click('#dashboard-hosts a:nth-child(%d)' % m3_index)
         b.wait_js_cond('window.location.hash != "#/dashboard/list"')
@@ -135,8 +135,8 @@ class TestMultiMachine(MachineCase):
         b.wait_text_not("#system_information_hostname_button", "")
         b.switch_to_top()
         b.go("/dashboard/list")
-        b.enter_page("dashboard")
-        b.wait_attr('#dashboard-hosts a:nth-child(%d) img' % m3_index, 'src', 'images/server-small.png')
+        b.enter_page("/dashboard/list")
+        b.wait_attr('#dashboard-hosts a:nth-child(%d) img' % m3_index, 'src', '../shell/images/server-small.png')
 
         self.allow_restart_journal_messages()
         self.allow_journal_messages(".*: .* host key for server is not known: .*",
@@ -157,7 +157,7 @@ class TestMultiMachine(MachineCase):
         m2.needs_writable_usr()
         m2.execute("gunzip -c /usr/share/cockpit/system/terminal.min.html.gz | sed -e 's|</body>|magic-m2-token</body>|' > /usr/share/cockpit/system/terminal.html")
 
-        self.login_and_go("dashboard", href="dashboard/list", host=None)
+        self.login_and_go("/dashboard/list", host=None)
         inject_extras(b)
         add_machine(b, m2.address)
         wait_dashboard_addresses (b, [ "localhost", m2.address ])
@@ -189,7 +189,7 @@ class TestMultiMachine(MachineCase):
         # Add a machine
         self.login_and_go(None)
         b.go(dashboard_path)
-        b.enter_page("dashboard")
+        b.enter_page("/dashboard/list")
         inject_extras(b)
         add_machine(b, m2.address)
         wait_dashboard_addresses (b, [ "localhost", m2.address ])
@@ -199,7 +199,7 @@ class TestMultiMachine(MachineCase):
         m2_index = wm[m2.address]+1
 
         # Check the server image
-        b.wait_attr('#dashboard-hosts a:nth-child(%d) img' % m2_index, 'src', 'images/server-small.png')
+        b.wait_attr('#dashboard-hosts a:nth-child(%d) img' % m2_index, 'src', '../shell/images/server-small.png')
         b.switch_to_top()
 
         # Go to the path, remove the image
@@ -217,8 +217,8 @@ class TestMultiMachine(MachineCase):
         kill_user_admin(m2)
 
         b.wait_js_cond("window.location.hash == '#%s'" % dashboard_path)
-        b.switch_to_frame("localhost/shell/shell")
-        b.wait_attr('#dashboard-hosts a:nth-child(%d) img' % m2_index, 'src', 'images/server-error.png')
+        b.switch_to_frame("localhost/dashboard/list")
+        b.wait_attr('#dashboard-hosts a:nth-child(%d) img' % m2_index, 'src', '../shell/images/server-error.png')
         b.switch_to_top()
 
         # navigating there again will fail
@@ -228,8 +228,8 @@ class TestMultiMachine(MachineCase):
         # renable admin
         m2.execute('passwd -u admin')
         # wait for dashboard to load
-        b.enter_page("dashboard")
-        b.wait_attr('#dashboard-hosts a:nth-child(%d) img' % m2_index, 'src', 'images/server-error.png')
+        b.enter_page("/dashboard/list")
+        b.wait_attr('#dashboard-hosts a:nth-child(%d) img' % m2_index, 'src', '../shell/images/server-error.png')
         b.switch_to_top()
 
         # path should load
@@ -241,8 +241,8 @@ class TestMultiMachine(MachineCase):
 
         # Dashboard shows it as up
         b.go(dashboard_path)
-        b.enter_page("dashboard")
-        b.wait_attr('#dashboard-hosts a:nth-child(%d) img' % m2_index, 'src', 'images/server-small.png')
+        b.enter_page("/dashboard/list")
+        b.wait_attr('#dashboard-hosts a:nth-child(%d) img' % m2_index, 'src', '../shell/images/server-small.png')
         b.switch_to_top()
 
         # Bad host also bounces
@@ -269,7 +269,7 @@ class TestMultiMachine(MachineCase):
         # Add a machine
         self.login_and_go(None)
         b.go("/dashboard/list")
-        b.enter_page("dashboard")
+        b.enter_page("/dashboard/list")
         inject_extras(b)
         add_machine(b, m2.address)
         wait_dashboard_addresses (b, [ "localhost", m2.address ])
