+++ test/check-multi-machine-key
@@ -107,7 +107,7 @@ class TestMultiMachineKeyAuth(MachineCase):
         m1 = self.machine
         m2 = self.machine2
 
-        self.login_and_go("dashboard", href="dashboard/list", host=None)
+        self.login_and_go("/dashboard/list", host=None)
         inject_extras(b)
         wait_dashboard_addresses (b, [ "localhost" ])
 
@@ -135,10 +135,10 @@ class TestMultiMachineKeyAuth(MachineCase):
         m1.execute("chown -R admin:admin /home/admin/.ssh")
 
         # Server is marked as down.
-        b.login_and_go("dashboard", href="dashboard/list", host=None)
+        b.login_and_go("/dashboard/list", host=None)
         inject_extras(b)
         wait_dashboard_addresses (b, [ "localhost", m2.address ])
-        b.wait_attr('#dashboard-hosts a:nth-child(%d) img' % m2_index, 'src', 'images/server-error.png')
+        b.wait_attr('#dashboard-hosts a:nth-child(%d) img' % m2_index, 'src', '../shell/images/server-error.png')
 
         try:
             m1.execute("ps xa | grep ssh-agent | grep -v grep")
@@ -167,8 +167,8 @@ class TestMultiMachineKeyAuth(MachineCase):
         b.wait_text_not("#system_information_hostname_button", "")
         b.switch_to_top()
         b.go("/dashboard/list")
-        b.enter_page("dashboard")
-        b.wait_attr('#dashboard-hosts a:nth-child(%d) img' % m2_index, 'src', 'images/server-small.png')
+        b.enter_page("/dashboard/list")
+        b.wait_attr('#dashboard-hosts a:nth-child(%d) img' % m2_index, 'src', '../shell/images/server-small.png')
 
         # Logout
         b.logout()
